make tag deprecator private in registry and remove from coverage
